requirements.txt
numpy==1.24.2
paho-mqtt==1.6.1
